prompt_handler/requirements.txt
Flask>=2.2.0
python-dotenv>=1.0.0
flask-cors>=3.0.10 